python/voice-live-quickstarts/TelemetryQuickstart/requirements.txt
azure-ai-voicelive
opentelemetry-sdk
azure-core-tracing-opentelemetry
azure-monitor-opentelemetry
